Remove local backend fallback (#1303)
* Merge joberror.py, jobtimeouterror.py & qiskitbackendnotfounderror.py
