+++ qiskit/backends/exceptions.py
@@ -1,0 +1,51 @@
+# -*- coding: utf-8 -*-
+
+# Copyright 2018, IBM.
+#
+# This source code is licensed under the Apache License, Version 2.0 found in
+# the LICENSE.txt file in the root directory of this source tree.
+
+"""
+Exceptions for errors raised while handling backends and jobs.
+"""
+
+from qiskit import QISKitError
+
+
+class JobError(QISKitError):
+    """Base class for errors raised by jobs."""
+
+    def __init__(self, *message):
+        """Set the error message."""
+        super().__init__(*message)
+        self.message = ' '.join(message)
+
+    def __str__(self):
+        """Return the message."""
+        return repr(self.message)
+
+
+class JobTimeoutError(JobError):
+    """Base class for timeout errors raised by jobs."""
+
+    def __init__(self, *message):
+        """Set the error message."""
+        super().__init__(*message)
+        self.message = ' '.join(message)
+
+    def __str__(self):
+        """Return the message."""
+        return repr(self.message)
+
+
+class QiskitBackendNotFoundError(QISKitError):
+    """Base class for errors raised while looking up for a backend."""
+
+    def __init__(self, *message):
+        """Set the error message."""
+        super().__init__(*message)
+        self.message = ' '.join(message)
+
+    def __str__(self):
+        """Return the message."""
+        return repr(self.message)
