--- qiskit/backends/joberror.py
@@ -1,25 +1,0 @@
-# -*- coding: utf-8 -*-
-
-# Copyright 2017, IBM.
-#
-# This source code is licensed under the Apache License, Version 2.0 found in
-# the LICENSE.txt file in the root directory of this source tree.
-
-"""
-Exception for errors raised by jobs.
-"""
-
-from qiskit import QISKitError
-
-
-class JobError(QISKitError):
-    """Base class for errors raised by jobs."""
-
-    def __init__(self, *message):
-        """Set the error message."""
-        super().__init__(*message)
-        self.message = ' '.join(message)
-
-    def __str__(self):
-        """Return the message."""
-        return repr(self.message)
